feat: add jpa repositories, StationService

+++ src/main/java/com/example/backend_vkr/entities/BaseEntity.java
@@ -4,16 +4,16 @@ import jakarta.persistence.*;
 
 @MappedSuperclass
 public abstract class BaseEntity {
-    private int id;
+    private Long id;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    protected void setId(int id) {
+    protected void setId(Long id) {
         this.id = id;
     }
 }
